Began work on Dialog UI
Added Text Mesh Pro and began working on the dialog ui. Now text pops up when dialog function is triggered, however, it is currently permanent. Would like to add a way to make it non-permanent and want to do the Money UI. Also, made an adjustment to the hearts ui so it creates new elements instead of using predefined ones.